// src/app/modules/transaction/transaction.interface.ts
/* eslint-disable no-unused-vars */
import { Types } from 'mongoose';

export enum PayStatus {
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  REVERSED = 'REVERSED',
}

export enum InitiatedBy {
  USER = 'USER',
  AGENT = 'AGENT',
}

export enum PayType {
  ADD_MONEY = 'ADD_MONEY',
  WITHDRAW = 'WITHDRAW',
  SEND_MONEY = 'SEND_MONEY',
  CASH_IN = 'CASH_IN',
  CASH_OUT = 'CASH_OUT',
}
export interface ITransaction {
  type: PayType;
  amount: number;
  contact: number;
  senderId?: Types.ObjectId;
  receiverId?: Types.ObjectId;
  wallet: Types.ObjectId;
  initiatedBy: InitiatedBy;
  fee?: number;
  commission?: number;
  status: PayStatus;
}